Backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
django-cors-headers
python-dotenv
django-cloudinary-storage
djangorestframework-simplejwt
django-redis
redis
gunicorn
whitenoise
psycopg2-binary
dj-database-url
pytest
pytest-django
coverage
pillow
requests
numpy>=1.24.0,<2.0.0
torch>=2.0.1
torchvision>=0.15.2
ultralytics>=8.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
django-anymail[brevo]